refactor: EventHandler 클래스 디렉토리 변경
- EventHandler의 위치를 common내 별도의 event 디렉토리 하위로 변경
- Spring Security 관련 exceptionHandler는 filter 내부로 이동
