+++ core/core-api/src/main/java/cmc/mellyserver/config/async/AsyncExceptionHandler.java
@@ -1,22 +1,18 @@
 package cmc.mellyserver.config.async;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
 
 import java.lang.reflect.Method;
 
+@Slf4j
 public class AsyncExceptionHandler implements AsyncUncaughtExceptionHandler {
 
-    private final Logger log = LoggerFactory.getLogger(getClass());
 
     @Override
     public void handleUncaughtException(Throwable e, Method method, Object... params) {
-        if (e instanceof Exception) {
 
-        } else {
-            log.error("Exception : {}", e.getMessage(), e);
-        }
+
     }
 
 }
